handled errors and exception

+++ src/main/java/com/example/contactsservice/Service/ContactService.java
@@ -1,7 +1,9 @@
 package com.example.contactsservice.Service;
 
+import com.example.contactsservice.BadRequest;
 import com.example.contactsservice.DTO.ContactDTO;
 import com.example.contactsservice.Entity.ContactEntity;
+import com.example.contactsservice.NotFound;
 import com.example.contactsservice.Repository.ContactRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -12,15 +14,25 @@ public class ContactService {
     ContactRepository contactRepository;
 
     public ContactEntity createContact(ContactDTO contactDTO) {
+        validateContactDto(contactDTO);
         ContactEntity contact = getContact(contactDTO, new ContactEntity());
         return contactRepository.save(contact);
     }
 
-    public ContactEntity getContact(Long id){
-        return contactRepository.getById(id);
+    public ContactEntity getContact(Long id) {
+        ContactEntity contact = contactRepository.getById(id);
+        try {
+            if (null != contact.getFirstName()) {
+                return contact;
+            }
+            return null;
+        } catch (RuntimeException runtimeException) {
+            throw new NotFound("User doesn't exist");
+        }
     }
 
     public ContactEntity updateContact(ContactDTO contactDTO, Long id) {
+        validateContactDto(contactDTO);
         ContactEntity contact = getContact(id);
         contact.setFirstName(contactDTO.getFirstName());
         contact.setLastName(contactDTO.getLastName());
@@ -28,11 +40,14 @@ public class ContactService {
         return contactRepository.save(contact);
     }
 
-    public void deleteContact(Long id){
-
+    public void deleteContact(Long id) {
         ContactEntity contact = getContact(id);
+        if (null != contact) {
+            contactRepository.delete(contact);
+        } else {
+            throw new NotFound("User doesn't exist");
+        }
 
-        contactRepository.delete(contact);
     }
 
     private ContactEntity getContact(ContactDTO contactDTO, ContactEntity contact) {
@@ -40,5 +55,14 @@ public class ContactService {
         contact.setLastName(contactDTO.getLastName());
         contact.setPhone(contactDTO.getPhone());
         return contact;
+    }
+
+    private Boolean validateContactDto(ContactDTO contactDTO) {
+        if (contactDTO.getFirstName().length() < 3 || contactDTO.getLastName().length() < 3) {
+            throw new BadRequest("Name should be more than 3 characters");
+        } else if (String.valueOf(contactDTO.getPhone()).length() < 10) {
+            throw new BadRequest("Invalid phone");
+        }
+        return true;
     }
 }
